feat: add Google GenAI embedder components for document and text embeddings
