--- integrations/google_ai/src/haystack_integrations/components/embedders/google_ai/__init__.py
@@ -1,3 +1,0 @@
-from .google_embedder import GoogleAIDocumentEmbedder, GoogleAITextEmbedder
-
-__all__ = ["GoogleAIDocumentEmbedder", "GoogleAITextEmbedder"]
