+++ integrations/google_genai/src/haystack_integrations/components/embedders/google_genai/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-FileCopyrightText: 2023-present deepset GmbH <info@deepset.ai>
+#
+# SPDX-License-Identifier: Apache-2.0
+from .document_embedder import GoogleAIDocumentEmbedder
+from .text_embedder import GoogleAITextEmbedder
+
+__all__ = ["GoogleAIDocumentEmbedder", "GoogleAITextEmbedder"]
